src/Basic-chain/requirements.txt
openai>=1.0.0
python-dotenv==1.0.0
langchain>=0.3.0
langchain-openai>=0.3.0